Move CMKEventConsole helper next to call sites
